Algrothim_4th/src/chapter1_Foundation/ResizingArraryStack.java
package chapter1_Foundation;

import java.util.Iterator;

public class ResizingArraryStack<Item> implements Iterable<Item> {

    private Item[] a = (Item[]) new Object[1];
    private int N = 0;

    public boolean isEmpty() {
        return N == 0;
    }

    public int size() {
        return N;
    }


    private void Resize(int max) {
        Item[] temp = (Item[]) new Object[max];
        for (int i = 0; i < N; i++) {
            temp[i] = a[i];
        }
        a = temp;
    }

    public void Push(Item item) {
        if (N == a.length) {
            Resize(2 * a.length);
        }
        a[N++] = item;
    }

    public Item Pop() {
        Item item = a[--N];
        a[N] = null;
        if (N > 0 && N == a.length / 4) {
            Resize(a.length / 2);
        }
        return item;
    }

    public Iterator<Item> iterator() {
        return new ReverseArraryIterator();
    }

    private class ReverseArraryIterator implements Iterator<Item> {

        private int i = N;

        @Override
        public boolean hasNext() {
            return i > 0;
        }

        @Override
        public Item next() {
            return a[--i];
        }
    }
}